Specialize vault key derivation
